Move additional classes to udmi package.
