+++ pubber/src/main/java/daq/pubber/RandomBoolean.java
@@ -1,6 +1,7 @@
 package daq.pubber;
 
 import udmi.lib.AbstractPoint;
+import udmi.lib.BasicPoint;
 import udmi.schema.PointPointsetModel;
 import udmi.schema.RefDiscovery;
