+++ pubber/src/main/java/udmi/lib/BasicPoint.java
@@ -1,11 +1,10 @@
-package daq.pubber;
+package udmi.lib;
 
 import static com.google.udmi.util.GeneralUtils.deepCopy;
 import static com.google.udmi.util.GeneralUtils.getNow;
 import static com.google.udmi.util.GeneralUtils.isTrue;
 
 import java.util.Objects;
-import udmi.lib.AbstractPoint;
 import udmi.schema.Category;
 import udmi.schema.Entry;
 import udmi.schema.PointPointsetConfig;
