+++ pubber/src/main/java/udmi/util/CatchingScheduledThreadPoolExecutor.java
@@ -1,8 +1,5 @@
-package com.google.daq.mqtt.util;
+package udmi.util;
 
-import java.util.concurrent.CancellationException;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 
 /**
